content/canvas/test/test_2d.fillStyle.parse.system.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.fillStyle.parse.system</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');


ctx.fillStyle = '#f00';
ctx.fillStyle = 'ThreeDDarkShadow';
ok(ctx.fillStyle =~ /^#(?!(FF0000|ff0000|f00)$)/, "ctx.fillStyle =~ /^#(?!(FF0000|ff0000|f00)$)/"); // test that it's not red

if (!_deferred) SimpleTest.finish();
});
</script>
